It changes and rename a few files in order to meet out intentions with the app. It scopes all the related request do Hangman game do '/hangman', rename controller and action.
It renames controller file, its module definition, the action and the template to call.

It renames the page folder templates folder to hangman.

It renames page_view file to hangman_view and it's module definition.

It replace all the new_game template file content with new content.

It removes all the content inside header tag and the alerts.

It just a few renamed files.